Don't unpack delta chains while converting delta to whole object
If an object is currently a delta but we need to convert it into
a whole object we shouldn't be chasing the delta chain to get its
real type code.  We had the type code at the time we made the list
of objects to pack; so we save it in the ObjectToPack structure.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>