Defer parsing of the ObjectId while walking a PackIndex Iterator
This is a slight performance improvement for the PackReverseIndex
construction.  The only code that really cares about the ObjectId
from a PackIndex entry is test cases.  By avoiding construction
we can save some CPU time during the several passes we do to make
the reverse index data structure within PackWriter.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>