Teach tree iterators to iterate from the workspace root
This is needed to be able to do treewalks of projects that
have their repository in the project's parent directory.

We always filter out '.git' entries when building the entry list,
no matter where we are in the tree, because we assume that the
checks in the ResourceEntry and FileEntry constructors for a
'.git' folder ensure that we never navigate into submodules.

Signed-off-by: Tor Arne Vestbø <torarnv@gmail.com>
Signed-off-by: Shawn O. Pearce <spearce@spearce.org>