Compute the author/committer name/email from git configuration
The author/committer name/email are retrieved from the system
configuration and the local and global git configurations.

When the name is not available, propose a default one by using the
system user name.

When the email is not available, propose a default one by
concatenating the user name and the host name.

The author name and email are used as the default value in the
author field while committing with the GUI.

Fix issue 47 (field Author don't fill)

Signed-off-by: Yann Simon <yann.simon.fr@gmail.com>
Signed-off-by: Shawn O. Pearce <spearce@spearce.org>