Append Signed-off-by text in the commit message editor
The user can see and edit the Signed-off-by line in the commit
dialog before committing.  Toggling the existing checkbox now
updates the message to add or remove the user's Signed-off-by
line, displaying the final message.

Updating the committer name and/or email also updates the
Signed-off-by line if it is present.

For new lines in the commit dialog, use Text.DELIMITER for
platform neutrality.

Signed-off-by: Yann Simon <yann.simon.fr@gmail.com>
Signed-off-by: Shawn O. Pearce <spearce@spearce.org>