Implement label decorations for folders and projects
The option "Re-decorate ancestors..." controls if parents of a
re-decorated resource also should be updated, for example to
to signal that the containing folder is now concidered dirty.

The option "Maximum number of levels..." controls how deep the
container decoration algorithm will recurse when trying to
determine the state (dirty, staged, etc.) of a container.

Signed-off-by: Tor Arne Vestbø <torarnv@gmail.com>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>