Arrange pack files in recency order to improve quick hits
Packs with newer modification dates are more likely to contain data
related to recent commits, and thus will have a better probability
of containing an object we care about when we are looking only at
recent history.

In some cases, this may permit JGit to avoid reading a very large
historical pack's index file into memory if the only data we ever
need is available in newer pack files.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>