Add the remote.name.mirror flag to RemoteConfig
This flag is currently only honored by "git push" to delete remote
refs if they are also gone on the source side.  It also causes the
default push spec to be "+refs/*:refs/*" if the user hasn't given
us any other push specs.

This change only implements the config file parsing for the mirror
flag, as there's still a chunk of code needed on the transport to
actually honor it.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>