Don't pass TagOpt to FetchProcess, get it from the Transport
We are already passing the transport down.  Its just easier to get
the tagopt off the transport than it is to pass it as an argument
to the constructor.

In the future as we add more options into the Transport we will also
need to obtain those in the FetchProcess, so directly querying the
Transport will be more consistent.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>