Fix sorting of thin packs in PackWriter
We must retain the TOPO and COMMIT_TIME_DESC rules when we add in
BOUNDARY for a thin pack, otherwise the resulting thin pack won't
match the expected ordering.

Its a non-critical error to sort the pack wrong, but it may cause the
client to skip around the pack data more frequently during access.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>