Require javadoc on protected methods in jgit core
In general protected methods are part of a public API since they grant
access to code outside of the package that declares the method or field.

A few protected fields and methods do not have javadocs, but it turns
out all but one are not used outside the package that declares them so
we can simply make them package private. This doesn't mean they aren't
supposed to be an API, but we can upgrade them if necessary.

A few places are actually used outside the package and are valid uses. For
those appropriate javadocs are added.

Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>
Signed-off-by: Shawn O. Pearce <spearce@spearce.org>