Split refspecs on the last : not the first
Recent discussion about a potential ':%branchname' syntax for
remote branches led rise to the question on #git of whether or
not this could be used on the lhs of a push refspec, such as
"+:%master:master".

Testing shows git-core permits this, as it must be splitting the
refspec with strrchr(), splitting along the last : in the string.
Make JGit match the split semantics.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>