Simplify the PackWriter.preparePack() API
It has always bothered me that we had two boolean "mode" parameters
as part of the preparePack method call, when most other options were
set by standard setter methods.  In all current callers (except the
unit tests) the ignoreMissingUninteresting was always set to true,
and in many of the callers, packthin was false.  Moving these to
setters with a reasonable default simplifies callers considerably.

Signed-off-by: Shawn O. Pearce <sop@google.com>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>