Addition of source and javadoc plugins to maven script
When JGIT jar is bundled with projects, it often needed
to inspect JGIT sources in order to look for documentation
and implementation. The sources are also very useful during
debugging. The patch adds source plugin that builds source jar,
so the matching version of the source.jar could be easily
bundled. The javadoc plugin generates ready to use archive with
javadoc.

Signed-off-by: Constantine Plotnikov <constantine.plotnikov@gmail.com>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>