Refactor EclipseSshSessionFactory to extend SshConfigSessionFactory
Keeping the EclipseSshSessionFactory in sync with our OpenSSH compliant
DefaultSshSessionFactory was a problem; e.g. a few commits ago we had
support for StrictHostKeyChecking added, but it was missed in the fork
we use under Eclipse.

Constantine Plotnikov's recent patch pointed out we should clean up
the code and make EclipseSshSessionFactory use the same code as the
other JGit clients use.  This makes it possible by refactoring the
way we create the Session and obtain a handle on the JSch.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>