Removed possibility to change stderr for ssh sessions
The current implementation allowed to change stderr for the
ssh sessions. However this functionality is broken. It is used
only by GitSshTransport and that class expects a very specific
behavior from this class. For example toString() method should
return the entire content of the stream. So only implementation
from SshConfigSessionFactory would have worked anyway. Returning
System.err (as was suggested by javadoc) comment would have broken
existing functionality. This patch makes this functionality
explicitly private.

If this functionality is to be reopened, this additional behavior
should be documented and there should be additional lifecycle
control, since the user streams will be interested to know when
stream will be no more used in order to release resoources.

Signed-off-by: Constantine Plotnikov <constantine.plotnikov@gmail.com>
Signed-off-by: Shawn O. Pearce <spearce@spearce.org>