Add tests to increase coverage of WindowCache
These tests run the WindowCache in different configurations with known
inputs, forcing it through different code paths as the cache populates
with entries, and potentially overflows.

The list of objects was derived from `git verify-pack -v`, run over all
of the pack files we install by default into a RepositoryTestCase.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>